# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  mode: string;
  /** @nullable */
  jabker?: string | null;
  /** @nullable */
  jenjang?: string | null;
  phase: string;
  createdAt: string;
}

export interface ConversationInput {
  title: string;
  mode: string;
  jabker?: string;
  jenjang?: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  mode: string;
  /** @nullable */
  jabker?: string | null;
  /** @nullable */
  jenjang?: string | null;
  phase: string;
  createdAt: string;
  messages: Message[];
}

export interface MessageInput {
  content: string;
}

export interface ExumGenerateInput {
  conversationId: number;
}

export interface ExumResult {
  content: string;
  conversationId: number;
}

export interface ApiError {
  error: string;
}