# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: chat
    description: Gustafta PKB chatbot operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /chat/conversations:
    get:
      operationId: listConversations
      tags: [chat]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [chat]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /chat/conversations/{id}:
    get:
      operationId: getConversation
      tags: [chat]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteConversation
      tags: [chat]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /chat/conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message and receive AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /chat/generate-exum:
    post:
      operationId: generateExum
      tags: [chat]
      summary: Generate Executive Summary document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExumGenerateInput"
      responses:
        "200":
          description: Generated Executive Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExumResult"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Conversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        mode:
          type: string
        jabker:
          type: ["string", "null"]
        jenjang:
          type: ["string", "null"]
        phase:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - mode
        - phase
        - createdAt
    ConversationInput:
      type: object
      properties:
        title:
          type: string
        mode:
          type: string
        jabker:
          type: string
        jenjang:
          type: string
      required:
        - title
        - mode
    ConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        mode:
          type: string
        jabker:
          type: ["string", "null"]
        jenjang:
          type: ["string", "null"]
        phase:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required:
        - id
        - title
        - mode
        - phase
        - createdAt
        - messages
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    MessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    ExumGenerateInput:
      type: object
      properties:
        conversationId:
          type: integer
      required:
        - conversationId
    ExumResult:
      type: object
      properties:
        content:
          type: string
        conversationId:
          type: integer
      required:
        - content
        - conversationId
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error