# lib/api-zod/src/generated/types/conversationWithMessages.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from './message';

export interface ConversationWithMessages {
  id: number;
  title: string;
  mode: string;
  /** @nullable */
  jabker?: string | null;
  /** @nullable */
  jenjang?: string | null;
  phase: string;
  createdAt: Date;
  messages: Message[];
}